Removal of redundant system code
The _health.py, _hide_my_tracks.py and _zipper.py have been moved to Logicytics.py, this is to make it easier to maintain
