+++ CODE/Logicytics.py
@@ -1,13 +1,13 @@
 from __future__ import annotations
 
+import datetime
+import hashlib
+import shutil
 import threading
+from datetime import datetime
 from typing import Any
 
 from __lib_class import *
-from _health import backup, update
-from _hide_my_tracks import attempt_hide
-from _zipper import Zip
-
 
 """
 This python script is the main entry point for the tool called Logicytics.
@@ -26,6 +26,221 @@ Here's a high-level overview of what the script does:
 The script appears to be designed to be highly configurable and modular, 
 with many options and flags that can be used to customize its behavior.
 """
+
+
+class Zip:
+    """
+    A class to handle zipping files, generating SHA256 hashes, and moving files.
+
+    Methods:
+        __get_files_to_zip(path: str) -> list:
+            Returns a list of files to be zipped, excluding certain file types and names.
+
+        __create_zip_file(path: str, files: list, filename: str):
+            Creates a zip file from the given list of files.
+
+        __remove_files(path: str, files: list):
+            Removes the specified files from the given path.
+
+        __generate_sha256_hash(filename: str) -> str:
+            Generates a SHA256 hash for the specified zip file.
+
+        __write_hash_to_file(filename: str, sha256_hash: str):
+            Writes the SHA256 hash to a file.
+
+        __move_files(filename: str):
+            Moves the zip file and its hash file to designated directories.
+
+        and_hash(self, path: str, name: str, flag: str) -> tuple | str:
+            Zips files, generates a SHA256 hash, and moves the files.
+    """
+
+    @staticmethod
+    def __get_files_to_zip(path: str) -> list:
+        """
+        Returns a list of files and directories to be zipped, excluding certain file types and names.
+
+        Args:
+            path (str): The directory path to search for files.
+
+        Returns:
+            list: A list of file and directory names to be zipped.
+        """
+        excluded_extensions = (".py", ".exe", ".bat", ".ps1")
+        excluded_prefixes = ("config.", "SysInternal_Suite", "__pycache__")
+
+        return [
+            f for f in os.listdir(path)
+            if not f.endswith(excluded_extensions) and not f.startswith(excluded_prefixes)
+        ]
+
+    @staticmethod
+    def __create_zip_file(path: str, files: list, filename: str):
+        """
+        Creates a zip file from the given list of files.
+
+        Args:
+            path (str): The directory path containing the files.
+            files (list): A list of file names to be zipped.
+            filename (str): The name of the output zip file.
+
+        Returns:
+            None
+        """
+        with zipfile.ZipFile(f"{filename}.zip", "w") as zip_file:
+            for file in files:
+                if os.path.isdir(os.path.join(path, file)):
+                    for root, _, files in os.walk(os.path.join(path, file)):
+                        for f in files:
+                            zip_file.write(os.path.join(root, f), os.path.relpath(os.path.join(root, f), path))
+                else:
+                    zip_file.write(os.path.join(path, file))
+
+    @staticmethod
+    def __remove_files(path: str, files: list):
+        """
+        Removes the specified files from the given path.
+
+        Args:
+            path (str): The directory path containing the files.
+            files (list): A list of file names to be removed.
+
+        Returns:
+            None or str: Returns an error message if an exception occurs.
+        """
+        for file in files:
+            try:
+                shutil.rmtree(os.path.join(path, file))
+            except OSError:
+                os.remove(os.path.join(path, file))
+            except Exception as e:
+                return f"Error: {e}"
+
+    @staticmethod
+    def __generate_sha256_hash(filename: str) -> str:
+        """
+        Generates a SHA256 hash for the specified zip file.
+
+        Args:
+            filename (str): The name of the zip file.
+
+        Returns:
+            str: The SHA256 hash of the zip file.
+        """
+        with open(f"{filename}.zip", "rb") as zip_file:
+            zip_data = zip_file.read()
+        return hashlib.sha256(zip_data).hexdigest()
+
+    @staticmethod
+    def __write_hash_to_file(filename: str, sha256_hash: str):
+        """
+        Writes the SHA256 hash to a file.
+
+        Args:
+            filename (str): The name of the hash file.
+            sha256_hash (str): The SHA256 hash to be written.
+
+        Returns:
+            None
+        """
+        with open(f"{filename}.hash", "w") as hash_file:
+            hash_file.write(sha256_hash)
+
+    @staticmethod
+    def __move_files(filename: str):
+        """
+        Moves the zip file and its hash file to designated directories.
+
+        Args:
+            filename (str): The name of the files to be moved.
+
+        Returns:
+            None
+        """
+        shutil.move(f"{filename}.zip", "../ACCESS/DATA/Zip")
+        shutil.move(f"{filename}.hash", "../ACCESS/DATA/Hashes")
+
+    def and_hash(self, path: str, name: str, flag: str) -> tuple | str:
+        """
+        Zips files, generates a SHA256 hash, and moves the files.
+
+        Args:
+            path (str): The directory path containing the files.
+            name (str): The base name for the output files.
+            flag (str): A flag to be included in the output file names.
+
+        Returns:
+            tuple or str: A tuple containing success messages or an error message.
+        """
+        today = datetime.now()
+        filename = f"Logicytics_{name}_{flag}_{today.strftime('%Y-%m-%d_%H-%M-%S')}"
+        files_to_zip = self.__get_files_to_zip(path)
+        self.__create_zip_file(path, files_to_zip, filename)
+        check = self.__remove_files(path, files_to_zip)
+        if isinstance(check, str):
+            return check
+        else:
+            sha256_hash = self.__generate_sha256_hash(filename)
+            self.__write_hash_to_file(filename, sha256_hash)
+            self.__move_files(filename)
+            return (
+                f"Zip file moved to ../ACCESS/DATA/Zip/{filename}.zip",
+                f"SHA256 Hash file moved to ../ACCESS/DATA/Hashes/{filename}.hash",
+            )
+
+
+class Health:
+    @staticmethod
+    def backup(directory: str, name: str) -> None:
+        """
+        Creates a backup of a specified directory by zipping its contents and moving it to a designated backup location.
+
+        Args:
+            directory (str): The path to the directory to be backed up.
+            name (str): The name of the backup file.
+
+        Returns:
+            None
+        """
+        # Check if backup exists, delete it if so
+        if os.path.exists(f"../ACCESS/BACKUP/{name}.zip"):
+            os.remove(f"../ACCESS/BACKUP/{name}.zip")
+
+        # Zip the directory and move it to the backup location
+        with zipfile.ZipFile(f"{name}.zip", "w") as zip_file:
+            for root, dirs, files in os.walk(directory):
+                for file in files:
+                    file_path = os.path.join(root, file)
+                    relative_path = os.path.relpath(str(file_path), start=os.getcwd())
+                    zip_file.write(str(file_path), arcname=relative_path)
+
+        shutil.move(f"{name}.zip", "../ACCESS/BACKUP")
+
+    @staticmethod
+    def update() -> tuple[str, str]:
+        """
+        Updates the repository by pulling the latest changes from the remote repository.
+
+        This function navigates to the parent directory, pulls the latest changes using Git,
+        and then returns to the current working directory.
+
+        Returns:
+            str: The output from the git pull command.
+        """
+        # Check if git command is available
+        if subprocess.run(["git", "--version"], capture_output=True).returncode != 0:
+            return "Git is not installed or not available in the PATH.", "error"
+
+        # Check if the project is a git repository
+        if not os.path.exists(os.path.join(os.getcwd(), ".git")):
+            return "This project is not a git repository. The update flag uses git.", "error"
+
+        current_dir = os.getcwd()
+        parent_dir = os.path.dirname(current_dir)
+        os.chdir(parent_dir)
+        output = subprocess.run(["git", "pull"], capture_output=True).stdout.decode()
+        os.chdir(current_dir)
+        return output, "info"
 
 
 def get_flags():
@@ -80,7 +295,7 @@ def handle_special_actions():
 
     if action == "update":
         log.info("Updating...")
-        message, log_type = update()
+        message, log_type = Health.update()
         log.string(message, log_type)
         log.info("Update complete!")
         input("Press Enter to exit...")
@@ -97,9 +312,9 @@ def handle_special_actions():
 
     if action == "backup":
         log.info("Backing up...")
-        backup(".", "Default_Backup")
+        Health.backup(".", "Default_Backup")
         log.debug("Backup complete -> CODE dir")
-        backup("../MODS", "Mods_Backup")
+        Health.backup("../MODS", "Mods_Backup")
         log.debug("Backup complete -> MODS dir")
         log.info("Backup complete!")
         input("Press Enter to exit...")
@@ -192,6 +407,24 @@ def generate_execution_list(actions: str) -> list | list[str] | list[str | Any]:
 
     log.debug(f"The following will be executed: {execution_list}")
     return execution_list
+
+
+def attempt_hide():
+    """
+    Attempts to delete Windows event logs from the current day.
+
+    Parameters:
+    None
+
+    Returns:
+    None
+    """
+    today = datetime.date.today()
+    log_path = r"C:\Windows\System32\winevt\Logs"
+
+    for file in os.listdir(log_path):
+        if file.endswith(".evtx") and file.startswith(today.strftime("%Y-%m-%d")):
+            subprocess.run(f'del "{os.path.join(log_path, file)}"', shell=False)
 
 
 def execute_scripts():
